fixed app params definition
