+++ src/main/webclient/src/App.tsx
@@ -1,9 +1,20 @@
 import React from "react";
 import "./App.css";
 import "@mantine/core/styles.css";
-import { AppMain } from "./components/app-main/app-main";
+import { AppMain, IAppMainProps } from "./components/app-main/app-main";
 
 export default function App() {
-
-  return <AppMain />;
+  const appMainProps: IAppMainProps = {
+    ui: {
+      colors: {
+        white: "#FFFFFF",
+        black: "#000000",
+        primary: "#22543D",
+        secondary: "#9AE6B4",
+        accent: "#CBD5E0",
+        background: "#F4F3F0",
+      },
+    },
+  };
+  return <AppMain {...appMainProps} />;
 }
