+++ src/main/webclient/src/components/app-main/app-main.tsx
@@ -5,23 +5,9 @@ import { ProjectColorsDictionary } from "../../types/projectTypes";
 export interface IAppMainProps {
   ui: IUIProps
 }
-
 export interface IUIProps {
   colors: ProjectColorsDictionary;
 }
-
-const appMainProps: IAppMainProps = {
-  ui: {
-    colors: {
-      white: "#FFFFFF",
-      black: "#000000",
-      primary: "#22543D",
-      secondary: "#9AE6B4",
-      accent: "#CBD5E0",
-      background: "#F4F3F0",
-    },
-  },
-};
 
 export const AppMain = ({...appMainProps}: IAppMainProps) => {
   return (
